Introduction of Novelty
When a creature enters into a tile, it attempts to add that tile to its
NoveltyChain with a corresponding expiration_frame. When that
expiration_frame is reached by the world generation counter, that tile
is removed from the NoveltyChain. If the tile is revisted, the
expiration_frame for that tile is updated. If a creature does not
explore a novel tile (a tile that isn't within the NoveltyChain) within
a given amount of time, the creature will be terminated. Children
inherit the NoveltyChain of their parent.

With this implementation, creatures seem to inhabit most areas of the
map that are inhabitable. The introduction of the NoveltyChain keeps
their population somewhat low as too many creatures in one area result
in tiles that continually have their expiration_frame updated.

There is a parameter that allows a NEW NoveltyChain to be passed onto
the child under a low probability, but this requires a significant
amount of experimenting with the other parameters of the environment.

Creatures also receive a very small benefit of Novelty when they kill
another creature. No idea how impactful this is.

+++ GridWorld/Tile.pde
@@ -79,7 +79,7 @@
                 shouldRedraw = false;
             //}
         }
-        void debug_draw(color c, int d) {
+        void debug_draw(color c, float d) {
             fill(c);
             rect(x-(w*d), y-(w*d), w*d*2, w*d*2);
         }
